RDKCOM-2533: RDKBDEV-667: use explicit 32bit data type in ANSC_IPV4_ADDRESS etc
Reason for change:
An IPv4 address is always 32 bits. Using "unsigned long" to hold it
leads to compiler warnings etc (and then various unnecessary casts
etc in the code to avoid them). Using the correct sized data type is
a better solution.
Test Procedure: Sanity.
Risks: None.

(cherry picked from commit 8a03c0f28a22789f9342d26b2561dddf7bb1ccba)

+++ source/util_api/http/include/http_wcso_interface.h
@@ -384,7 +384,7 @@ typedef  ANSC_STATUS
 (*PFN_HTTPWCSO_CONNECT_PEER)
     (
         ANSC_HANDLE                 hThisObject,
-        ULONG                       ipAddr
+        uint32_t                    ipAddr
     );
 
 typedef  ANSC_STATUS
